Extract default market session config
